DataProcessor/requirements.txt
tqdm>=4.62.3
pandas>=1.3.4
beautifulsoup4>=4.1.0
google-cloud-storage>=2.1.0
google-cloud-bigquery>=2.34.0